dead

+++ src/RequestHandlers/Post/upload.js
@@ -1,13 +1,9 @@
-const client = require("ipfs-http-client");
 const fs = require('mz/fs');
 const path = require("path");
 const crypto = require("../../Encryption");
 const Database = require("../../Database");
 
-const ipfs = client("ipfs.infura.io", 5001, {protocol: "https"});
 const MAX_SIZE = 52428800;
-
-const algo = "aes-256-ctr"
 
 module.exports = async (req, res) => {
     if(req.user == null){
@@ -63,28 +59,26 @@ module.exports = async (req, res) => {
         });
     }
 
-    //Encrypt the file with algo "aes-256-ctr";
+    if(req.file.originalname.length >= 250){
+        return res.status(500).json({
+            error: "File Name too Long."
+        })
+    }
 
-    console.log("Path: ", req.file.path);
     const saveAs = `${req.file.path}-enc`;
 
     try{
         const now = parseInt(Date.now() / 1000);
         const iv = await crypto.encryptFile(crypto.md5(`${req.user.password}@${now}`), fs.createReadStream(req.file.path), fs.createWriteStream(saveAs));
         await fs.unlink(req.file.path);
-        const dbResult = await Database.addFile(req.user.id, iv, now, req.file.mimetype);
-        
-        const files = await ipfs.add(fs.readFileSync(saveAs), {
-            progress: prog => console.log(`received: ${prog}`)
-        });
-
-        await fs.unlink(saveAs);
-        await Database.uploadedFile(dbResult.id, files[0].hash, parseInt(Date.now() / 1000));
-        console.log(files[0].hash);
-        return res.json({
-            hash: files[0].hash,
-            name: req.file.originalname
-        });
+        const dbResult = await Database.uploadFile(req.user.id, iv, now, req.file.mimetype, path.basename(saveAs), path.extname(req.file.originalname), req.file.originalname);
+        if(dbResult.affectedRows >= 1){
+            console.log(`File ID: ${dbResult.id}`);
+            return res.json({
+                id: dbResult.id,
+                name: req.file.originalname
+            });
+        }else throw new Error("File did not uploaded to database.")
     }catch(e){
         console.log(e);
         return res.status(500).json({
